RegExp literal not instance of RegExp fixed with new env
extends jest-env-node to use all that presetting instead and avoid setting global.RegExp as this is what breaks instanceof. Problem is that Object in VM context are different distinct objects.
CF https://github.com/jsdom/jsdom/compare/7.0.1...7.0.2 where jsdom removed that Date and RegExp setting in this for the same reason